pyproject.toml
[tool.poetry]
name = "doggo"
version = "0.1.0"
description = ""
authors = ["Meoyoug <eodyd426@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.dev.dependencies]
django = ">=5.0.1,<6.0.0"
djangorestframework = ">=3.14.0,<4.0.0"
djangorestframework-simplejwt = ">=5.3.1,<6.0.0"
mysqlclient = ">=2.2.4,<3.0.0"
drf-spectacular = ">=0.27.1,<0.28.0"
uwsgi = ">=2.0.24,<2.1.0"
django-crontab = ">=0.7.1,<0.8.0"
python-dotenv = ">=1.0.1,<2.0.0"
django-cors-headers = ">=4.3.1,<5.0.0"
django-allauth = ">=0.61.1,<1.0.0"
mypy = ">=1.8.0,<2.0.0"
black = ">=24.3.0,<25.0.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
python_version = 3.12
strict = true
